config : add dependancy for pid.0.bias on motion type
with the bias being dependant on motion type you can set
different values for the bias.

when extruding the temperature will drop, thus you need a
higher bias to compensate for this to keep the temperature level.
you can now make a bias depending on the value of the speed with
a lincurve, but for easy use, just look at the motion type